+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Changed
+- Evaluate `SendingAudioFailureConnectionHealthPolicy` only after getting connected i.e. after `audioVideoDidStart()` is called. This avoids false positive triggers of `sendingAudioFailed` at the start of a meeting session.
 
 ### Fixed